# swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "E-commerce API",
      "description": "API documentation for the E-commerce application"
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "schemes": ["http"],
    "paths": {
...
      "/products": {
        "get": {
          "summary": "Retrieve a list of products",
          "parameters": [
            {
              "in": "query",
              "name": "category",
              "required": false,
              "description": "ID of the category to filter products by",
              "type": "integer"
            }
          ],
          "responses": {
            "200": {
              "description": "A list of products",
              "schema": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "integer", "example": 1 },
                    "name": { "type": "string", "example": "Sample Product" },
                    "description": { "type": "string", "example": "This is a sample product" },
                    "price": { "type": "number", "format": "float", "example": 29.99 },
                    "stock": { "type": "integer", "example": 100 },
                    "category_id": { "type": "integer", "example": 1 }
                  }
                }
              }
            }
          }
        }
      },
      "/products/{productId}": {
        "get": {
          "summary": "Retrieve a specific product",
          "parameters": [
            {
              "in": "path",
              "name": "productId",
              "required": true,
              "description": "ID of the product to retrieve",
              "type": "integer"
            }
          ],
          "responses": {
            "200": {
              "description": "A product object",
              "schema": {
                "type": "object",
                "properties": {
                  "id": { "type": "integer", "example": 1 },
                  "name": { "type": "string", "example": "Sample Product" },
                  "description": { "type": "string", "example": "This is a sample product" },
                  "price": { "type": "number", "format": "float", "example": 29.99 },
                  "stock": { "type": "integer", "example": 100 },
                  "category_id": { "type": "integer", "example": 1 }
                }
              }
            },
            "404": {
              "description": "Product not found"
            }
          }
        }
      },
...
    }
  }